Changed Meta-movement navigation to Ctrl-arrow
Meta keys are a lot of "fun" in terminals so I would prefer not to use
them in key maps.  However, we can't transfer all the meta key maps into
ctrl maps because terminals often have no way of differentiating between
<BS> and <C-h>.  I think similar problems might happen with <C-j> too.

I didn't want to use ctrl-arrow initially because it will override WORDS
navigation, but it isn't a big deal since we still have W and B and most
of the time words navigation is all you need, and shift-arrow will still
do that (not in some terminals though!)

Another option I should consider is just remapping C-w to something
different.  The real problem is that it is just such a pain to type C-w
all the time.  For now though, I think I'll just stick with the
ctrl-arrows.