Made clearing search highlighting less buggy
<Esc><Esc> will now be converted to <Esc> when there is no highlighting
to clear.  This means that mistyping <Up> as <Esc><Up> will no longer
result in an "A" inserted on the line above, and instead have the more
desirable behavior of moving up.  With the previous mapping, I
frequently found myself typing <Esc> 3+ times to clear the highlighting
instead of twice, and then trying to move around, which resulted in the
undesirable behavior.