Even better search highlight clearing
91d1a0c didn't work on older versions of vim because v:hlsearch wasn't
defined.  This time we ONLY make <Esc> meaningful if it comes before
another <Esc> or as part of an arrow key or similar.  <Esc> by itself
will no longer beep, which is a good thing because simply having an
<Esc><Esc> mapping combined with notimeout makes that feedback unhelpful
because it will only happen at the keystroke after the one you need it.