Made base64 decoding work well cross-platform
I carefully used syntax that works the same in both bash and windows
batch to make the <Leader>b64 mapping work on windows.  You will still
need to create 'base64' and 'printf' programs for this to work, but
that's not too hard, especially if you don't worry about silly things
like POSIX compliance.

The parentheses are needed because on windows, vim essentially
transforms your command into (your command < temp_file) instead of (your
command) < temp_file like I think it does on other platforms.  Because
of this, only the printf would get the right stdin, so we need to add
the parentheses ourselves.  However, if we JUST do that, it breaks other
platforms because in bash it would get transformed into ((your command))
< temp_file which would evaluate as an arithmetic expression (see the
bash man page).  Because of this, we add a leading space before the
command.

I COULD just check to see if the user is running windows, but that's no
fun!