Fix spell checking in glsl and yaml
When syntax definitions do not contain any @Spell clusters, the default
is to spell-check EVERYTHING.  To fix this, we could disable syntax
spell-checking (see :syn-spell), but we want comments to get spell
checked so we add the @Spell cluster to glsl and yaml's "Comment" rule
which confines syntax highlighting to comments.  In yaml, this is a
little more complicated/hacky because yaml doesn't define a "Comment"
cluster.