Use a submodule for tomorrow-theme
I like to have all my plugins as submodules.  Previously I didn't use it
for tomorrow-theme because of the way
https://github.com/chriskempson/tomorrow-theme is organized.  I recently
found that @chriskempson also has a repo just for vim, but unfortunately
it is out of date and has diverged significantly from the main
tomorrow-theme.  It is also marked as deprecated in favor of base16.
However, I prefer Tomorrow-Night to base16-tomorrow-night.  base16 is
also rather large, containing a lot of themes I just don't care about.
Maybe sometime in the future I'll switch to base16, but for now I
decided to use @jsit's repo which merges the main tomorrow-theme repo
with the vim one.