Text objects for lines and files
These text objects don't do anything particularly interesting.  For the
most part they are just things that my fingers think should exist when
I'm vimming.  However, the inner versions of the text objects can be
quite useful.  For example, yil is not the same as yy because yil
ignores any surrounding whitespace, including the newline.

As a side note, writing good text objects can be surprisingly difficult.
The builtin text objects extend visual mode selections.  To mimic this
behavior for lines, you need to know if you are at the top of the
selection or not so that you know which direction to move in to complete
the line.  Note that we don't even try to support visual block mode
here.  Luckily, the builtin text objects don't work any better with
visual block mode from what I can tell.

At some point, I think it would be neat if I could create a "syntax
region" text object that selects everything in the current region.  That
would be really nice with the gw command: gwir on a comment would
intelligently format it.